+++ app/src/main/java/za/co/opsmobile/coindispense/dispense/action/CostFetchedAction.java
@@ -1,5 +1,7 @@
 package za.co.opsmobile.coindispense.dispense.action;
 
+import java.math.BigDecimal;
+
 import za.co.opsmobile.coindispense.dispense.store.DispenseStoreActions;
 import za.co.opsmobile.coindispense.framework.action.StoreAction;
 
@@ -7,9 +9,9 @@ import za.co.opsmobile.coindispense.framework.action.StoreAction;
  * Created by Daniel Oosthuizen on 2015/09/07.
  */
 public class CostFetchedAction implements StoreAction<DispenseStoreActions> {
-    private final float cost;
+    private final BigDecimal cost;
 
-    public CostFetchedAction(float cost) {
+    public CostFetchedAction(BigDecimal cost) {
         this.cost = cost;
     }
