+++ model/src/main/java/za/co/opsmobile/coindispense/CalculateChangeRequest.java
@@ -1,25 +1,27 @@
 package za.co.opsmobile.coindispense;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Daniel Oosthuizen on 2015/09/13.
  */
 public class CalculateChangeRequest {
-    private float cost;
-    private float payment;
+    private BigDecimal cost;
+    private BigDecimal payment;
 
-    public float getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
 
-    public void setCost(float cost) {
+    public void setCost(BigDecimal cost) {
         this.cost = cost;
     }
 
-    public float getPayment() {
+    public BigDecimal getPayment() {
         return payment;
     }
 
-    public void setPayment(float payment) {
+    public void setPayment(BigDecimal payment) {
         this.payment = payment;
     }
 }
