+++ model/src/main/java/za/co/opsmobile/coindispense/ChangeResult.java
@@ -1,26 +1,27 @@
 package za.co.opsmobile.coindispense;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 /**
  * Created by Daniel Oosthuizen on 2015/09/13.
  */
 public class ChangeResult {
-    private HashMap<Float, Integer> payments;
+    private HashMap<BigDecimal, Integer> payments;
 
     public ChangeResult() {
         payments = new HashMap<>();
     }
 
-    public HashMap<Float, Integer> getPayments() {
+    public HashMap<BigDecimal, Integer> getPayments() {
         return payments;
     }
 
-    public void setPayments(HashMap<Float, Integer> payments) {
+    public void setPayments(HashMap<BigDecimal, Integer> payments) {
         this.payments = payments;
     }
 
-    public void addPayment(float payment) {
+    public void addPayment(BigDecimal payment) {
         if (payments.containsKey(payment)) {
             payments.put(payment, payments.get(payment) + 1);
         } else {
